Home work 1
