+++ src/main/java/lesson1/HomeWork1/Cat.java
@@ -1,9 +1,38 @@
 package lesson1.HomeWork1;
 
-public class Cat extends Participants{
+public class Cat implements Participants {
+    private String name;
+    private int runLimit;
+    private int jumpLimit;
 
+    public Cat(String name, int runLimit, int jumpLimit) {
+        this.name = name;
+        this.runLimit = runLimit;
+        this.jumpLimit = jumpLimit;
+    }
 
-    public Cat(String name, int jumpLimit, int runLimit) {
-        super(name, jumpLimit, runLimit);
+    @Override
+    public void jump() {
+        System.out.println(name + " перепрыгнул");
+    }
+
+    @Override
+    public void run() {
+        System.out.println(name + " пробежал");
+    }
+
+    @Override
+    public boolean obstacle(Obstacle obstacle) {
+        if (obstacle.getTredmill() <= runLimit && obstacle.getWall() == 0) {
+            return true;
+        } else if (obstacle.getWall() <= jumpLimit && obstacle.getTredmill() == 0) {
+            return true;
+        }
+        return false;
+    }
+
+    @Override
+    public String getName() {
+        return name;
     }
 }
