+++ src/main/java/lesson1/HomeWork1/Jumpable.java
@@ -1,6 +1,5 @@
 package lesson1.HomeWork1;
 
-
 public interface Jumpable {
 
     default void jump() {
