+++ src/main/java/lesson1/HomeWork1/Obstacle.java
@@ -1,7 +1,7 @@
 package lesson1.HomeWork1;
 
 public interface Obstacle {
-int getTredmill();
-int getWall();
+    int getTredmill();
 
+    int getWall();
 }
