+++ src/main/java/lesson1/HomeWork1/Robot.java
@@ -1,9 +1,38 @@
 package lesson1.HomeWork1;
 
-public class Robot extends Participants{
+public class Robot implements Participants {
+    private String name;
+    private int runLimit;
+    private int jumpLimit;
 
+    public Robot(String name, int runLimit, int jumpLimit) {
+        this.name = name;
+        this.runLimit = runLimit;
+        this.jumpLimit = jumpLimit;
+    }
 
-    public Robot(String name, int jumpLimit, int runLimit) {
-        super(name, jumpLimit, runLimit);
+    @Override
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public void jump() {
+        System.out.println(name + " перепрыгнул");
+    }
+
+    @Override
+    public void run() {
+        System.out.println(name + " пробежал");
+    }
+
+    @Override
+    public boolean obstacle(Obstacle obstacle) {
+        if (obstacle.getTredmill() <= runLimit && obstacle.getWall() == 0) {
+            return true;
+        } else if (obstacle.getWall() <= jumpLimit && obstacle.getTredmill() == 0) {
+            return true;
+        }
+        return false;
     }
 }
