--- src/main/java/lesson1/Wall.java
@@ -1,20 +1,0 @@
-package lesson1;
-
-import lesson1.interfaces.Obstacles;
-
-public class Wall implements Obstacles {
-    private int height;
-
-    public Wall(int height) {
-        this.height = height;
-    }
-
-    @Override
-    public int getLength() {
-        return 0;
-    }
-
-    public int getHeight() {
-        return height;
-    }
-}
