--- src/main/java/lesson1/interfaces/Obstacles.java
@@ -1,7 +1,0 @@
-package lesson1.interfaces;
-
-public interface Obstacles {
-
-    int getLength();
-    int getHeight();
-}
