--- src/main/java/lesson1/interfaces/Participants.java
@@ -1,6 +1,0 @@
-package lesson1.interfaces;
-
-public interface Participants {
-    boolean obstacle(Obstacles obstacles);
-    String getName();
-}
